Correctly display carry weight limit.
Former-commit-id: 15ae2ef14b0af9274ba806811f9cc9f9d680e098

+++ client/src/include/player.h
@@ -142,7 +142,7 @@ typedef struct Player_Struct
 	sint32 container_tag;
 
 	/** Player's weight limit. */
-	uint32 weight_limit;
+	double weight_limit;
 
 	/** Are we a DM? */
 	int dm;
